Updated Forms
Moved some elements around, included an upload image button.

Last thing to figure out: How to move elements up a little bit in the left form.

How to change image text to an actual image for when it is uploaded.

Have information stay there until it is changed once again.

Last two are dealt with database.